Reorganize AwsChunked/conent_encoding logic back to single module
