+++ aws/rust-runtime/aws-runtime/src/content_encoding.rs
@@ -4,15 +4,17 @@
  */
 
 use bytes::{Bytes, BytesMut};
-use http_02x::{HeaderMap, HeaderValue};
-use http_body_04x::{Body, SizeHint};
 use pin_project_lite::pin_project;
 
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
 const CRLF: &str = "\r\n";
+const CRLF_RAW: &[u8] = b"\r\n";
+
 const CHUNK_TERMINATOR: &str = "0\r\n";
+const CHUNK_TERMINATOR_RAW: &[u8] = b"0\r\n";
+
 const TRAILER_SEPARATOR: &[u8] = b":";
 
 /// Content encoding header value constants
@@ -136,70 +138,9 @@ impl<Inner> AwsChunkedBody<Inner> {
     }
 }
 
-fn get_unsigned_chunk_bytes_length(payload_length: u64) -> u64 {
-    let hex_repr_len = int_log16(payload_length);
-    hex_repr_len + CRLF.len() as u64 + payload_length + CRLF.len() as u64
-}
-
-/// Writes trailers out into a `string` and then converts that `String` to a `Bytes` before
-/// returning.
-///
-/// - Trailer names are separated by a single colon only, no space.
-/// - Trailer names with multiple values will be written out one line per value, with the name
-///   appearing on each line.
-fn trailers_as_aws_chunked_bytes(
-    trailer_map: Option<HeaderMap>,
-    estimated_length: u64,
-) -> BytesMut {
-    if let Some(trailer_map) = trailer_map {
-        let mut current_header_name = None;
-        let mut trailers = BytesMut::with_capacity(estimated_length.try_into().unwrap_or_default());
-
-        for (header_name, header_value) in trailer_map.into_iter() {
-            // When a header has multiple values, the name only comes up in iteration the first time
-            // we see it. Therefore, we need to keep track of the last name we saw and fall back to
-            // it when `header_name == None`.
-            current_header_name = header_name.or(current_header_name);
-
-            // In practice, this will always exist, but `if let` is nicer than unwrap
-            if let Some(header_name) = current_header_name.as_ref() {
-                trailers.extend_from_slice(header_name.as_ref());
-                trailers.extend_from_slice(TRAILER_SEPARATOR);
-                trailers.extend_from_slice(header_value.as_bytes());
-                trailers.extend_from_slice(CRLF.as_bytes());
-            }
-        }
-
-        trailers
-    } else {
-        BytesMut::new()
-    }
-}
-
-/// Given an optional `HeaderMap`, calculate the total number of bytes required to represent the
-/// `HeaderMap`. If no `HeaderMap` is given as input, return 0.
-///
-/// - Trailer names are separated by a single colon only, no space.
-/// - Trailer names with multiple values will be written out one line per value, with the name
-///   appearing on each line.
-fn total_rendered_length_of_trailers(trailer_map: Option<&HeaderMap>) -> u64 {
-    match trailer_map {
-        Some(trailer_map) => trailer_map
-            .iter()
-            .map(|(trailer_name, trailer_value)| {
-                trailer_name.as_str().len()
-                    + TRAILER_SEPARATOR.len()
-                    + trailer_value.len()
-                    + CRLF.len()
-            })
-            .sum::<usize>() as u64,
-        None => 0,
-    }
-}
-
-impl<Inner> Body for AwsChunkedBody<Inner>
+impl<Inner> http_body_04x::Body for AwsChunkedBody<Inner>
 where
-    Inner: Body<Data = Bytes, Error = aws_smithy_types::body::Error>,
+    Inner: http_body_04x::Body<Data = Bytes, Error = aws_smithy_types::body::Error>,
 {
     type Data = Bytes;
     type Error = aws_smithy_types::body::Error;
@@ -257,7 +198,8 @@ where
                 return match this.inner.poll_trailers(cx) {
                     Poll::Ready(Ok(trailers)) => {
                         *this.state = AwsChunkedBodyState::Closed;
-                        let expected_length = total_rendered_length_of_trailers(trailers.as_ref());
+                        let expected_length =
+                            http_02x_utils::total_rendered_length_of_trailers(trailers.as_ref());
                         let actual_length = this.options.total_trailer_length();
 
                         if expected_length != actual_length {
@@ -269,8 +211,10 @@ where
                             return Poll::Ready(Some(Err(err)));
                         }
 
-                        let mut trailers =
-                            trailers_as_aws_chunked_bytes(trailers, actual_length + 1);
+                        let mut trailers = http_02x_utils::trailers_as_aws_chunked_bytes(
+                            trailers,
+                            actual_length + 1,
+                        );
                         // Insert the final CRLF to close the body
                         trailers.extend_from_slice(CRLF.as_bytes());
 
@@ -287,7 +231,7 @@ where
     fn poll_trailers(
         self: Pin<&mut Self>,
         _cx: &mut Context<'_>,
-    ) -> Poll<Result<Option<HeaderMap<HeaderValue>>, Self::Error>> {
+    ) -> Poll<Result<Option<http_02x::HeaderMap<http_02x::HeaderValue>>, Self::Error>> {
         // Trailers were already appended to the body because of the content encoding scheme
         Poll::Ready(Ok(None))
     }
@@ -296,8 +240,296 @@ where
         self.state == AwsChunkedBodyState::Closed
     }
 
-    fn size_hint(&self) -> SizeHint {
-        SizeHint::with_exact(self.encoded_length())
+    fn size_hint(&self) -> http_body_04x::SizeHint {
+        http_body_04x::SizeHint::with_exact(self.encoded_length())
+    }
+}
+
+/// Utility functions to help with the [http_body_04x::Body] trait implementation
+mod http_02x_utils {
+    use super::{CRLF, TRAILER_SEPARATOR};
+    use bytes::BytesMut;
+    use http_02x::HeaderMap;
+
+    /// Writes trailers out into a `string` and then converts that `String` to a `Bytes` before
+    /// returning.
+    ///
+    /// - Trailer names are separated by a single colon only, no space.
+    /// - Trailer names with multiple values will be written out one line per value, with the name
+    ///   appearing on each line.
+    pub(super) fn trailers_as_aws_chunked_bytes(
+        trailer_map: Option<HeaderMap>,
+        estimated_length: u64,
+    ) -> BytesMut {
+        if let Some(trailer_map) = trailer_map {
+            let mut current_header_name = None;
+            let mut trailers =
+                BytesMut::with_capacity(estimated_length.try_into().unwrap_or_default());
+
+            for (header_name, header_value) in trailer_map.into_iter() {
+                // When a header has multiple values, the name only comes up in iteration the first time
+                // we see it. Therefore, we need to keep track of the last name we saw and fall back to
+                // it when `header_name == None`.
+                current_header_name = header_name.or(current_header_name);
+
+                // In practice, this will always exist, but `if let` is nicer than unwrap
+                if let Some(header_name) = current_header_name.as_ref() {
+                    trailers.extend_from_slice(header_name.as_ref());
+                    trailers.extend_from_slice(TRAILER_SEPARATOR);
+                    trailers.extend_from_slice(header_value.as_bytes());
+                    trailers.extend_from_slice(CRLF.as_bytes());
+                }
+            }
+
+            trailers
+        } else {
+            BytesMut::new()
+        }
+    }
+
+    /// Given an optional `HeaderMap`, calculate the total number of bytes required to represent the
+    /// `HeaderMap`. If no `HeaderMap` is given as input, return 0.
+    ///
+    /// - Trailer names are separated by a single colon only, no space.
+    /// - Trailer names with multiple values will be written out one line per value, with the name
+    ///   appearing on each line.
+    pub(super) fn total_rendered_length_of_trailers(trailer_map: Option<&HeaderMap>) -> u64 {
+        match trailer_map {
+            Some(trailer_map) => trailer_map
+                .iter()
+                .map(|(trailer_name, trailer_value)| {
+                    trailer_name.as_str().len()
+                        + TRAILER_SEPARATOR.len()
+                        + trailer_value.len()
+                        + CRLF.len()
+                })
+                .sum::<usize>() as u64,
+            None => 0,
+        }
+    }
+}
+
+const UNREACHABLE_STATES: &str = "These states already short circuited";
+
+///
+impl<Inner> http_body_1x::Body for AwsChunkedBody<Inner>
+where
+    Inner: http_body_1x::Body<Data = Bytes, Error = aws_smithy_types::body::Error>,
+{
+    type Data = Bytes;
+    type Error = aws_smithy_types::body::Error;
+
+    fn is_end_stream(&self) -> bool {
+        self.state == AwsChunkedBodyState::Closed
+    }
+
+    fn size_hint(&self) -> http_body_1x::SizeHint {
+        http_body_1x::SizeHint::with_exact(self.encoded_length())
+    }
+
+    fn poll_frame(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<Option<Result<http_body_1x::Frame<Self::Data>, Self::Error>>> {
+        tracing::trace!(state = ?self.state, "polling AwsChunkedBody");
+        let mut this = self.project();
+
+        // Both `WritingChunkSize` and `Closed` states short circuit without polling the inner body
+
+        // Initial setup, we do not poll the inner body here
+        if *this.state == AwsChunkedBodyState::WritingChunkSize {
+            if this.options.stream_length == 0 {
+                // If the stream is empty, we skip to writing trailers after writing the CHUNK_TERMINATOR.
+                tracing::trace!("stream is empty, writing chunk terminator");
+                let frame = http_body_1x::Frame::data(Bytes::from(CHUNK_TERMINATOR));
+                *this.state = AwsChunkedBodyState::WritingTrailers;
+                return Poll::Ready(Some(Ok(frame)));
+            } else {
+                // A chunk must be prefixed by chunk size in hexadecimal
+                let chunk_size = format!(
+                    "{:X?}{}",
+                    this.options.stream_length,
+                    std::str::from_utf8(CRLF_RAW).unwrap()
+                );
+                tracing::trace!(%chunk_size, "writing chunk size");
+                let chunk_size = http_body_1x::Frame::data(Bytes::from(chunk_size));
+                *this.state = AwsChunkedBodyState::WritingChunk;
+                return Poll::Ready(Some(Ok(chunk_size)));
+            }
+        }
+
+        // Polled after completion
+        if *this.state == AwsChunkedBodyState::Closed {
+            return Poll::Ready(None);
+        }
+
+        // For all other states we must poll the inner body
+        let maybe_frame = this.inner.poll_frame(cx);
+        tracing::trace!(poll_state = ?maybe_frame, "Polling InnerBody");
+
+        match maybe_frame {
+            Poll::Ready(Some(Ok(frame))) => match *this.state {
+                // Both data chunks and trailers are written as Frame::data so we treat these states similarly
+                // Importantly we cannot know that the body data of the InnerBody is exhausted until we see a
+                // trailer frame or a Poll::Ready(None)
+                AwsChunkedBodyState::WritingChunk => {
+                    if frame.is_data() {
+                        let data = frame.data_ref().expect("Data frame has data");
+                        tracing::trace!(len = data.len(), "Writing chunk data");
+                        *this.inner_body_bytes_read_so_far += data.len();
+                        Poll::Ready(Some(Ok(frame)))
+                    } else {
+                        tracing::trace!(
+                            "No more chunk data, writing CRLF + CHUNK_TERMINATOR to end the data, and the first trailer frame"
+                        );
+                        *this.state = AwsChunkedBodyState::WritingTrailers;
+                        let trailers = frame.trailers_ref();
+
+                        // NOTE: there is a subtle logic bug here (which is present in the http-02x implementation as well)
+                        // The check for this error assumes that all trailers will come in a single trailer frame. Currently
+                        // I believe this will always be the case since the only thing we send trailers for in AwsChunked is
+                        // streaming checksums and that is a single trailer value. But it might not always be true. We should
+                        // fix this bug when we update the behavior here to match the actual spec.
+                        // The fix probably looks like returning Poll::Pending while we buffer all of the trailers and then
+                        // comparing the actual length to the expected length before returning a final frame containing all
+                        // of the trailers.
+                        let actual_length: u64 =
+                            http_1x_utils::total_rendered_length_of_trailers(trailers);
+                        let expected_length = this.options.total_trailer_length();
+                        if expected_length != actual_length {
+                            let err =
+                                Box::new(AwsChunkedBodyError::ReportedTrailerLengthMismatch {
+                                    actual: actual_length,
+                                    expected: expected_length,
+                                });
+                            return Poll::Ready(Some(Err(err)));
+                        }
+
+                        // Capacity = actual_length (in case all of the trailers specified in  come in AwsChunkedBodyOptions
+                        // come in the first trailer frame which is going to be the case most of the time in practice) + 7
+                        // (2 + 3) for the initial CRLF + CHUNK_TERMINATOR to end the chunked data + 2 for the final CRLF
+                        // ending the trailers section.
+                        let mut buf = BytesMut::with_capacity(actual_length as usize + 7);
+                        // End the final data chunk
+                        buf.extend_from_slice(&[CRLF_RAW, CHUNK_TERMINATOR_RAW].concat());
+
+                        // We transform the trailers into raw bytes. We can't write them with Frame::trailers
+                        // since we must include the CRLF + CHUNK_TERMINATOR that end the body and the CRLFs
+                        // after each trailer, so we write them as Frame::data
+                        let trailers = http_1x_utils::trailers_as_aws_chunked_bytes(trailers, buf);
+                        Poll::Ready(Some(Ok(http_body_1x::Frame::data(trailers.into()))))
+                    }
+                }
+                AwsChunkedBodyState::WritingTrailers => {
+                    let trailers = frame.trailers_ref();
+                    let actual_length: u64 =
+                        http_1x_utils::total_rendered_length_of_trailers(trailers);
+                    let buf = BytesMut::with_capacity(actual_length as usize + 7);
+                    let trailers = http_1x_utils::trailers_as_aws_chunked_bytes(trailers, buf);
+                    Poll::Ready(Some(Ok(http_body_1x::Frame::data(trailers.into()))))
+                }
+                AwsChunkedBodyState::Closed | AwsChunkedBodyState::WritingChunkSize => {
+                    unreachable!("{}", UNREACHABLE_STATES)
+                }
+            },
+            // InnerBody data exhausted, add finalizing bytes depending on current state
+            Poll::Ready(None) => {
+                let trailers = match *this.state {
+                    AwsChunkedBodyState::WritingChunk => {
+                        let actual_stream_length = *this.inner_body_bytes_read_so_far as u64;
+                        let expected_stream_length = this.options.stream_length;
+                        if actual_stream_length != expected_stream_length {
+                            let err = Box::new(AwsChunkedBodyError::StreamLengthMismatch {
+                                actual: actual_stream_length,
+                                expected: expected_stream_length,
+                            });
+                            return Poll::Ready(Some(Err(err)));
+                        };
+                        let mut trailers = BytesMut::with_capacity(7);
+                        trailers.extend_from_slice(
+                            &[CRLF_RAW, CHUNK_TERMINATOR_RAW, CRLF_RAW].concat(),
+                        );
+                        trailers
+                    }
+                    AwsChunkedBodyState::WritingTrailers => {
+                        let mut trailers = BytesMut::with_capacity(2);
+                        trailers.extend_from_slice(CRLF_RAW);
+                        trailers
+                    }
+                    AwsChunkedBodyState::Closed | AwsChunkedBodyState::WritingChunkSize => {
+                        unreachable!("{}", UNREACHABLE_STATES)
+                    }
+                };
+
+                let frame = http_body_1x::Frame::data(trailers.into());
+                *this.state = AwsChunkedBodyState::Closed;
+                Poll::Ready(Some(Ok(frame)))
+            }
+            // Passthrough states
+            Poll::Pending => Poll::Pending,
+            Poll::Ready(Some(Err(e))) => Poll::Ready(Some(Err(e))),
+        }
+    }
+}
+/// Utility functions to help with the [http_body_1x::Body] trait implementation
+mod http_1x_utils {
+    use super::{CRLF_RAW, TRAILER_SEPARATOR};
+    use bytes::BytesMut;
+    use http_1x::{HeaderMap, HeaderName};
+
+    /// Writes trailers out into a `string` and then converts that `String` to a `Bytes` before
+    /// returning.
+    ///
+    /// - Trailer names are separated by a single colon only, no space.
+    /// - Trailer names with multiple values will be written out one line per value, with the name
+    ///   appearing on each line.
+    pub(super) fn trailers_as_aws_chunked_bytes(
+        trailer_map: Option<&HeaderMap>,
+        mut buffer: BytesMut,
+    ) -> BytesMut {
+        if let Some(trailer_map) = trailer_map {
+            let mut current_header_name: Option<HeaderName> = None;
+
+            for (header_name, header_value) in trailer_map.clone().into_iter() {
+                // When a header has multiple values, the name only comes up in iteration the first time
+                // we see it. Therefore, we need to keep track of the last name we saw and fall back to
+                // it when `header_name == None`.
+                current_header_name = header_name.or(current_header_name);
+
+                // In practice, this will always exist, but `if let` is nicer than unwrap
+                if let Some(header_name) = current_header_name.as_ref() {
+                    buffer.extend_from_slice(header_name.as_ref());
+                    buffer.extend_from_slice(TRAILER_SEPARATOR);
+                    buffer.extend_from_slice(header_value.as_bytes());
+                    buffer.extend_from_slice(CRLF_RAW);
+                }
+            }
+
+            buffer
+        } else {
+            buffer
+        }
+    }
+
+    /// Given an optional `HeaderMap`, calculate the total number of bytes required to represent the
+    /// `HeaderMap`. If no `HeaderMap` is given as input, return 0.
+    ///
+    /// - Trailer names are separated by a single colon only, no space.
+    /// - Trailer names with multiple values will be written out one line per value, with the name
+    ///   appearing on each line.
+    pub(super) fn total_rendered_length_of_trailers(trailer_map: Option<&HeaderMap>) -> u64 {
+        match trailer_map {
+            Some(trailer_map) => trailer_map
+                .iter()
+                .map(|(trailer_name, trailer_value)| {
+                    trailer_name.as_str().len()
+                        + TRAILER_SEPARATOR.len()
+                        + trailer_value.len()
+                        + CRLF_RAW.len()
+                })
+                .sum::<usize>() as u64,
+            None => 0,
+        }
     }
 }
 
@@ -347,92 +579,221 @@ where
     len
 }
 
+fn get_unsigned_chunk_bytes_length(payload_length: u64) -> u64 {
+    let hex_repr_len = int_log16(payload_length);
+    hex_repr_len + CRLF.len() as u64 + payload_length + CRLF.len() as u64
+}
+
 #[cfg(test)]
 mod tests {
-    use super::{
-        total_rendered_length_of_trailers, trailers_as_aws_chunked_bytes, AwsChunkedBody,
-        AwsChunkedBodyOptions, CHUNK_TERMINATOR, CRLF,
-    };
 
-    use aws_smithy_types::body::SdkBody;
-    use bytes::{Buf, Bytes};
-    use bytes_utils::SegmentedBuf;
-    use http_02x::{HeaderMap, HeaderValue};
-    use http_body_04x::{Body, SizeHint};
-    use pin_project_lite::pin_project;
+    #[cfg(test)]
+    mod http_02x_tests {
+        use super::super::{
+            http_02x_utils::{total_rendered_length_of_trailers, trailers_as_aws_chunked_bytes},
+            AwsChunkedBody, AwsChunkedBodyOptions, CHUNK_TERMINATOR, CRLF,
+        };
 
-    use std::io::Read;
-    use std::pin::Pin;
-    use std::task::{Context, Poll};
-    use std::time::Duration;
+        use aws_smithy_types::body::SdkBody;
+        use bytes::{Buf, Bytes};
+        use bytes_utils::SegmentedBuf;
+        use http_02x::{HeaderMap, HeaderValue};
+        use http_body_04x::{Body, SizeHint};
+        use pin_project_lite::pin_project;
 
-    pin_project! {
-        struct SputteringBody {
-            parts: Vec<Option<Bytes>>,
-            cursor: usize,
-            delay_in_millis: u64,
-        }
-    }
+        use std::io::Read;
+        use std::pin::Pin;
+        use std::task::{Context, Poll};
+        use std::time::Duration;
 
-    impl SputteringBody {
-        fn len(&self) -> usize {
-            self.parts.iter().flatten().map(|b| b.len()).sum()
-        }
-    }
-
-    impl Body for SputteringBody {
-        type Data = Bytes;
-        type Error = aws_smithy_types::body::Error;
-
-        fn poll_data(
-            self: Pin<&mut Self>,
-            cx: &mut Context<'_>,
-        ) -> Poll<Option<Result<Self::Data, Self::Error>>> {
-            if self.cursor == self.parts.len() {
-                return Poll::Ready(None);
-            }
-
-            let this = self.project();
-            let delay_in_millis = *this.delay_in_millis;
-            let next_part = this.parts.get_mut(*this.cursor).unwrap().take();
-
-            match next_part {
-                None => {
-                    *this.cursor += 1;
-                    let waker = cx.waker().clone();
-                    tokio::spawn(async move {
-                        tokio::time::sleep(Duration::from_millis(delay_in_millis)).await;
-                        waker.wake();
-                    });
-                    Poll::Pending
-                }
-                Some(data) => {
-                    *this.cursor += 1;
-                    Poll::Ready(Some(Ok(data)))
-                }
+        pin_project! {
+            struct SputteringBody {
+                parts: Vec<Option<Bytes>>,
+                cursor: usize,
+                delay_in_millis: u64,
             }
         }
 
-        fn poll_trailers(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-        ) -> Poll<Result<Option<HeaderMap<HeaderValue>>, Self::Error>> {
-            Poll::Ready(Ok(None))
+        impl SputteringBody {
+            fn len(&self) -> usize {
+                self.parts.iter().flatten().map(|b| b.len()).sum()
+            }
         }
 
-        fn is_end_stream(&self) -> bool {
-            false
+        impl Body for SputteringBody {
+            type Data = Bytes;
+            type Error = aws_smithy_types::body::Error;
+
+            fn poll_data(
+                self: Pin<&mut Self>,
+                cx: &mut Context<'_>,
+            ) -> Poll<Option<Result<Self::Data, Self::Error>>> {
+                if self.cursor == self.parts.len() {
+                    return Poll::Ready(None);
+                }
+
+                let this = self.project();
+                let delay_in_millis = *this.delay_in_millis;
+                let next_part = this.parts.get_mut(*this.cursor).unwrap().take();
+
+                match next_part {
+                    None => {
+                        *this.cursor += 1;
+                        let waker = cx.waker().clone();
+                        tokio::spawn(async move {
+                            tokio::time::sleep(Duration::from_millis(delay_in_millis)).await;
+                            waker.wake();
+                        });
+                        Poll::Pending
+                    }
+                    Some(data) => {
+                        *this.cursor += 1;
+                        Poll::Ready(Some(Ok(data)))
+                    }
+                }
+            }
+
+            fn poll_trailers(
+                self: Pin<&mut Self>,
+                _cx: &mut Context<'_>,
+            ) -> Poll<Result<Option<HeaderMap<HeaderValue>>, Self::Error>> {
+                Poll::Ready(Ok(None))
+            }
+
+            fn is_end_stream(&self) -> bool {
+                false
+            }
+
+            fn size_hint(&self) -> SizeHint {
+                SizeHint::new()
+            }
         }
 
-        fn size_hint(&self) -> SizeHint {
-            SizeHint::new()
-        }
-    }
+        #[tokio::test]
+        async fn test_aws_chunked_encoding() {
+            let test_fut = async {
+                let input_str = "Hello world";
+                let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, Vec::new());
+                let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
 
-    #[tokio::test]
-    async fn test_aws_chunked_encoding() {
-        let test_fut = async {
+                let mut output = SegmentedBuf::new();
+                while let Some(buf) = body.data().await {
+                    output.push(buf.unwrap());
+                }
+
+                let mut actual_output = String::new();
+                output
+                    .reader()
+                    .read_to_string(&mut actual_output)
+                    .expect("Doesn't cause IO errors");
+
+                let expected_output = "B\r\nHello world\r\n0\r\n\r\n";
+
+                assert_eq!(expected_output, actual_output);
+                assert!(
+                    body.trailers()
+                        .await
+                        .expect("no errors occurred during trailer polling")
+                        .is_none(),
+                    "aws-chunked encoded bodies don't have normal HTTP trailers"
+                );
+
+                // You can insert a `tokio::time::sleep` here to verify the timeout works as intended
+            };
+
+            let timeout_duration = Duration::from_secs(3);
+            if tokio::time::timeout(timeout_duration, test_fut)
+                .await
+                .is_err()
+            {
+                panic!("test_aws_chunked_encoding timed out after {timeout_duration:?}");
+            }
+        }
+
+        #[tokio::test]
+        async fn test_aws_chunked_encoding_sputtering_body() {
+            let test_fut = async {
+                let input = SputteringBody {
+                    parts: vec![
+                        Some(Bytes::from_static(b"chunk 1, ")),
+                        None,
+                        Some(Bytes::from_static(b"chunk 2, ")),
+                        Some(Bytes::from_static(b"chunk 3, ")),
+                        None,
+                        None,
+                        Some(Bytes::from_static(b"chunk 4, ")),
+                        Some(Bytes::from_static(b"chunk 5, ")),
+                        Some(Bytes::from_static(b"chunk 6")),
+                    ],
+                    cursor: 0,
+                    delay_in_millis: 500,
+                };
+                let opts = AwsChunkedBodyOptions::new(input.len() as u64, Vec::new());
+                let mut body = AwsChunkedBody::new(input, opts);
+
+                let mut output = SegmentedBuf::new();
+                while let Some(buf) = body.data().await {
+                    output.push(buf.unwrap());
+                }
+
+                let mut actual_output = String::new();
+                output
+                    .reader()
+                    .read_to_string(&mut actual_output)
+                    .expect("Doesn't cause IO errors");
+
+                let expected_output =
+                    "34\r\nchunk 1, chunk 2, chunk 3, chunk 4, chunk 5, chunk 6\r\n0\r\n\r\n";
+
+                assert_eq!(expected_output, actual_output);
+                assert!(
+                    body.trailers()
+                        .await
+                        .expect("no errors occurred during trailer polling")
+                        .is_none(),
+                    "aws-chunked encoded bodies don't have normal HTTP trailers"
+                );
+            };
+
+            let timeout_duration = Duration::from_secs(3);
+            if tokio::time::timeout(timeout_duration, test_fut)
+                .await
+                .is_err()
+            {
+                panic!(
+                "test_aws_chunked_encoding_sputtering_body timed out after {timeout_duration:?}"
+            );
+            }
+        }
+
+        #[tokio::test]
+        #[should_panic = "called `Result::unwrap()` on an `Err` value: ReportedTrailerLengthMismatch { actual: 44, expected: 0 }"]
+        async fn test_aws_chunked_encoding_incorrect_trailer_length_panic() {
             let input_str = "Hello world";
+            // Test body has no trailers, so this length is incorrect and will trigger an assert panic
+            // When the panic occurs, it will actually expect a length of 44. This is because, when using
+            // aws-chunked encoding, each trailer will end with a CRLF which is 2 bytes long.
+            let wrong_trailer_len = 42;
+            let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, vec![wrong_trailer_len]);
+            let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
+
+            // We don't care about the body contents but we have to read it all before checking for trailers
+            while let Some(buf) = body.data().await {
+                drop(buf.unwrap());
+            }
+
+            assert!(
+                body.trailers()
+                    .await
+                    .expect("no errors occurred during trailer polling")
+                    .is_none(),
+                "aws-chunked encoded bodies don't have normal HTTP trailers"
+            );
+        }
+
+        #[tokio::test]
+        async fn test_aws_chunked_encoding_empty_body() {
+            let input_str = "";
             let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, Vec::new());
             let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
 
@@ -447,7 +808,7 @@ mod tests {
                 .read_to_string(&mut actual_output)
                 .expect("Doesn't cause IO errors");
 
-            let expected_output = "B\r\nHello world\r\n0\r\n\r\n";
+            let expected_output = [CHUNK_TERMINATOR, CRLF].concat();
 
             assert_eq!(expected_output, actual_output);
             assert!(
@@ -457,43 +818,227 @@ mod tests {
                     .is_none(),
                 "aws-chunked encoded bodies don't have normal HTTP trailers"
             );
+        }
 
-            // You can insert a `tokio::time::sleep` here to verify the timeout works as intended
-        };
+        #[tokio::test]
+        async fn test_total_rendered_length_of_trailers() {
+            let mut trailers = HeaderMap::new();
 
-        let timeout_duration = Duration::from_secs(3);
-        if tokio::time::timeout(timeout_duration, test_fut)
-            .await
-            .is_err()
-        {
-            panic!("test_aws_chunked_encoding timed out after {timeout_duration:?}");
+            trailers.insert("empty_value", HeaderValue::from_static(""));
+
+            trailers.insert("single_value", HeaderValue::from_static("value 1"));
+
+            trailers.insert("two_values", HeaderValue::from_static("value 1"));
+            trailers.append("two_values", HeaderValue::from_static("value 2"));
+
+            trailers.insert("three_values", HeaderValue::from_static("value 1"));
+            trailers.append("three_values", HeaderValue::from_static("value 2"));
+            trailers.append("three_values", HeaderValue::from_static("value 3"));
+
+            let trailers = Some(trailers);
+            let actual_length = total_rendered_length_of_trailers(trailers.as_ref());
+            let expected_length =
+                (trailers_as_aws_chunked_bytes(trailers, actual_length).len()) as u64;
+
+            assert_eq!(expected_length, actual_length);
+        }
+
+        #[tokio::test]
+        async fn test_total_rendered_length_of_empty_trailers() {
+            let trailers = Some(HeaderMap::new());
+            let actual_length = total_rendered_length_of_trailers(trailers.as_ref());
+            let expected_length =
+                (trailers_as_aws_chunked_bytes(trailers, actual_length).len()) as u64;
+
+            assert_eq!(expected_length, actual_length);
         }
     }
 
-    #[tokio::test]
-    async fn test_aws_chunked_encoding_sputtering_body() {
-        let test_fut = async {
-            let input = SputteringBody {
-                parts: vec![
-                    Some(Bytes::from_static(b"chunk 1, ")),
-                    None,
-                    Some(Bytes::from_static(b"chunk 2, ")),
-                    Some(Bytes::from_static(b"chunk 3, ")),
-                    None,
-                    None,
-                    Some(Bytes::from_static(b"chunk 4, ")),
-                    Some(Bytes::from_static(b"chunk 5, ")),
-                    Some(Bytes::from_static(b"chunk 6")),
-                ],
-                cursor: 0,
-                delay_in_millis: 500,
+    #[cfg(test)]
+    mod http_1x_tests {
+        use super::super::{
+            http_1x_utils::{total_rendered_length_of_trailers, trailers_as_aws_chunked_bytes},
+            AwsChunkedBody, AwsChunkedBodyOptions, CHUNK_TERMINATOR_RAW, CRLF_RAW,
+        };
+
+        use aws_smithy_types::body::SdkBody;
+        use bytes::{Buf, Bytes, BytesMut};
+        use bytes_utils::SegmentedBuf;
+        use http_1x::{HeaderMap, HeaderValue};
+        use http_body_1x::{Body, Frame, SizeHint};
+        use http_body_util::BodyExt;
+        use pin_project_lite::pin_project;
+
+        use std::io::Read;
+        use std::pin::Pin;
+        use std::task::{Context, Poll};
+        use std::time::Duration;
+
+        pin_project! {
+            struct SputteringBody {
+                parts: Vec<Option<Bytes>>,
+                cursor: usize,
+                delay_in_millis: u64,
+            }
+        }
+
+        impl SputteringBody {
+            fn len(&self) -> usize {
+                self.parts.iter().flatten().map(|b| b.len()).sum()
+            }
+        }
+
+        impl Body for SputteringBody {
+            type Data = Bytes;
+            type Error = aws_smithy_types::body::Error;
+
+            fn poll_frame(
+                self: Pin<&mut Self>,
+                cx: &mut Context<'_>,
+            ) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
+                if self.cursor == self.parts.len() {
+                    return Poll::Ready(None);
+                }
+
+                let this = self.project();
+                let delay_in_millis = *this.delay_in_millis;
+                let next_part = this.parts.get_mut(*this.cursor).unwrap().take();
+
+                match next_part {
+                    None => {
+                        *this.cursor += 1;
+                        let waker = cx.waker().clone();
+                        tokio::spawn(async move {
+                            tokio::time::sleep(Duration::from_millis(delay_in_millis)).await;
+                            waker.wake();
+                        });
+                        Poll::Pending
+                    }
+                    Some(data) => {
+                        *this.cursor += 1;
+                        let frame = Frame::data(data);
+                        Poll::Ready(Some(Ok(frame)))
+                    }
+                }
+            }
+
+            fn is_end_stream(&self) -> bool {
+                false
+            }
+
+            fn size_hint(&self) -> SizeHint {
+                SizeHint::new()
+            }
+        }
+
+        #[tokio::test]
+        async fn test_aws_chunked_encoding() {
+            let test_fut = async {
+                let input_str = "Hello world";
+                let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, vec![]);
+                let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
+
+                let mut output = SegmentedBuf::new();
+                while let Some(Ok(buf)) = body.frame().await {
+                    output.push(buf.into_data().unwrap());
+                }
+
+                let mut actual_output = String::new();
+                output
+                    .reader()
+                    .read_to_string(&mut actual_output)
+                    .expect("Doesn't cause IO errors");
+
+                let expected_output = "B\r\nHello world\r\n0\r\n\r\n";
+
+                assert_eq!(expected_output, actual_output);
+
+                // You can insert a `tokio::time::sleep` here to verify the timeout works as intended
             };
-            let opts = AwsChunkedBodyOptions::new(input.len() as u64, Vec::new());
-            let mut body = AwsChunkedBody::new(input, opts);
+
+            let timeout_duration = Duration::from_secs(3);
+            if tokio::time::timeout(timeout_duration, test_fut)
+                .await
+                .is_err()
+            {
+                panic!("test_aws_chunked_encoding timed out after {timeout_duration:?}");
+            }
+        }
+
+        #[tokio::test]
+        async fn test_aws_chunked_encoding_sputtering_body() {
+            let test_fut = async {
+                let input = SputteringBody {
+                    parts: vec![
+                        Some(Bytes::from_static(b"chunk 1, ")),
+                        None,
+                        Some(Bytes::from_static(b"chunk 2, ")),
+                        Some(Bytes::from_static(b"chunk 3, ")),
+                        None,
+                        None,
+                        Some(Bytes::from_static(b"chunk 4, ")),
+                        Some(Bytes::from_static(b"chunk 5, ")),
+                        Some(Bytes::from_static(b"chunk 6")),
+                    ],
+                    cursor: 0,
+                    delay_in_millis: 500,
+                };
+                let opts = AwsChunkedBodyOptions::new(input.len() as u64, vec![]);
+                let mut body = AwsChunkedBody::new(input, opts);
+
+                let mut output = SegmentedBuf::new();
+                while let Some(Ok(buf)) = body.frame().await {
+                    output.push(buf.into_data().unwrap());
+                }
+
+                let mut actual_output = String::new();
+                output
+                    .reader()
+                    .read_to_string(&mut actual_output)
+                    .expect("Doesn't cause IO errors");
+
+                let expected_output =
+                    "34\r\nchunk 1, chunk 2, chunk 3, chunk 4, chunk 5, chunk 6\r\n0\r\n\r\n";
+
+                assert_eq!(expected_output, actual_output);
+            };
+
+            let timeout_duration = Duration::from_secs(3);
+            if tokio::time::timeout(timeout_duration, test_fut)
+                .await
+                .is_err()
+            {
+                panic!(
+                "test_aws_chunked_encoding_sputtering_body timed out after {timeout_duration:?}"
+            );
+            }
+        }
+
+        #[tokio::test]
+        async fn test_aws_chunked_encoding_incorrect_trailer_length_panic() {
+            let input_str = "Hello world";
+            // Test body has no trailers, so this length is incorrect and will trigger an assert panic
+            // When the panic occurs, it will actually expect a length of 44. This is because, when using
+            // aws-chunked encoding, each trailer will end with a CRLF which is 2 bytes long.
+            let wrong_trailer_len = 42;
+            let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, vec![wrong_trailer_len]);
+            let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
+
+            // We don't care about the body contents but we have to read it all before checking for trailers
+            while let Some(Ok(frame)) = body.frame().await {
+                assert!(!frame.is_trailers());
+            }
+        }
+
+        #[tokio::test]
+        async fn test_aws_chunked_encoding_empty_body() {
+            let input_str = "";
+            let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, vec![]);
+            let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
 
             let mut output = SegmentedBuf::new();
-            while let Some(buf) = body.data().await {
-                output.push(buf.unwrap());
+            while let Some(Ok(frame)) = body.frame().await {
+                output.push(frame.into_data().unwrap());
             }
 
             let mut actual_output = String::new();
@@ -502,112 +1047,45 @@ mod tests {
                 .read_to_string(&mut actual_output)
                 .expect("Doesn't cause IO errors");
 
-            let expected_output =
-                "34\r\nchunk 1, chunk 2, chunk 3, chunk 4, chunk 5, chunk 6\r\n0\r\n\r\n";
+            let actual_output = std::str::from_utf8(actual_output.as_bytes()).unwrap();
+            let expected_output = [CHUNK_TERMINATOR_RAW, CRLF_RAW].concat();
+            let expected_output = std::str::from_utf8(&expected_output).unwrap();
 
             assert_eq!(expected_output, actual_output);
-            assert!(
-                body.trailers()
-                    .await
-                    .expect("no errors occurred during trailer polling")
-                    .is_none(),
-                "aws-chunked encoded bodies don't have normal HTTP trailers"
-            );
-        };
-
-        let timeout_duration = Duration::from_secs(3);
-        if tokio::time::timeout(timeout_duration, test_fut)
-            .await
-            .is_err()
-        {
-            panic!(
-                "test_aws_chunked_encoding_sputtering_body timed out after {timeout_duration:?}"
-            );
-        }
-    }
-
-    #[tokio::test]
-    #[should_panic = "called `Result::unwrap()` on an `Err` value: ReportedTrailerLengthMismatch { actual: 44, expected: 0 }"]
-    async fn test_aws_chunked_encoding_incorrect_trailer_length_panic() {
-        let input_str = "Hello world";
-        // Test body has no trailers, so this length is incorrect and will trigger an assert panic
-        // When the panic occurs, it will actually expect a length of 44. This is because, when using
-        // aws-chunked encoding, each trailer will end with a CRLF which is 2 bytes long.
-        let wrong_trailer_len = 42;
-        let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, vec![wrong_trailer_len]);
-        let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
-
-        // We don't care about the body contents but we have to read it all before checking for trailers
-        while let Some(buf) = body.data().await {
-            drop(buf.unwrap());
         }
 
-        assert!(
-            body.trailers()
-                .await
-                .expect("no errors occurred during trailer polling")
-                .is_none(),
-            "aws-chunked encoded bodies don't have normal HTTP trailers"
-        );
-    }
+        #[tokio::test]
+        async fn test_total_rendered_length_of_trailers() {
+            let mut trailers = HeaderMap::new();
 
-    #[tokio::test]
-    async fn test_aws_chunked_encoding_empty_body() {
-        let input_str = "";
-        let opts = AwsChunkedBodyOptions::new(input_str.len() as u64, Vec::new());
-        let mut body = AwsChunkedBody::new(SdkBody::from(input_str), opts);
+            trailers.insert("empty_value", HeaderValue::from_static(""));
 
-        let mut output = SegmentedBuf::new();
-        while let Some(buf) = body.data().await {
-            output.push(buf.unwrap());
+            trailers.insert("single_value", HeaderValue::from_static("value 1"));
+
+            trailers.insert("two_values", HeaderValue::from_static("value 1"));
+            trailers.append("two_values", HeaderValue::from_static("value 2"));
+
+            trailers.insert("three_values", HeaderValue::from_static("value 1"));
+            trailers.append("three_values", HeaderValue::from_static("value 2"));
+            trailers.append("three_values", HeaderValue::from_static("value 3"));
+
+            let trailers = Some(&trailers);
+            let actual_length = total_rendered_length_of_trailers(trailers);
+            let buf = BytesMut::with_capacity(actual_length as usize);
+            let expected_length = (trailers_as_aws_chunked_bytes(trailers, buf).len()) as u64;
+
+            assert_eq!(expected_length, actual_length);
         }
 
-        let mut actual_output = String::new();
-        output
-            .reader()
-            .read_to_string(&mut actual_output)
-            .expect("Doesn't cause IO errors");
+        #[tokio::test]
+        async fn test_total_rendered_length_of_empty_trailers() {
+            let header_map = HeaderMap::new();
+            let trailers = Some(&header_map);
+            let actual_length = total_rendered_length_of_trailers(trailers);
+            let buf = BytesMut::with_capacity(actual_length as usize);
+            let expected_length = (trailers_as_aws_chunked_bytes(trailers, buf).len()) as u64;
 
-        let expected_output = [CHUNK_TERMINATOR, CRLF].concat();
-
-        assert_eq!(expected_output, actual_output);
-        assert!(
-            body.trailers()
-                .await
-                .expect("no errors occurred during trailer polling")
-                .is_none(),
-            "aws-chunked encoded bodies don't have normal HTTP trailers"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_total_rendered_length_of_trailers() {
-        let mut trailers = HeaderMap::new();
-
-        trailers.insert("empty_value", HeaderValue::from_static(""));
-
-        trailers.insert("single_value", HeaderValue::from_static("value 1"));
-
-        trailers.insert("two_values", HeaderValue::from_static("value 1"));
-        trailers.append("two_values", HeaderValue::from_static("value 2"));
-
-        trailers.insert("three_values", HeaderValue::from_static("value 1"));
-        trailers.append("three_values", HeaderValue::from_static("value 2"));
-        trailers.append("three_values", HeaderValue::from_static("value 3"));
-
-        let trailers = Some(trailers);
-        let actual_length = total_rendered_length_of_trailers(trailers.as_ref());
-        let expected_length = (trailers_as_aws_chunked_bytes(trailers, actual_length).len()) as u64;
-
-        assert_eq!(expected_length, actual_length);
-    }
-
-    #[tokio::test]
-    async fn test_total_rendered_length_of_empty_trailers() {
-        let trailers = Some(HeaderMap::new());
-        let actual_length = total_rendered_length_of_trailers(trailers.as_ref());
-        let expected_length = (trailers_as_aws_chunked_bytes(trailers, actual_length).len()) as u64;
-
-        assert_eq!(expected_length, actual_length);
+            assert_eq!(expected_length, actual_length);
+        }
     }
 }
